Fix Webpack 5 build
Fix the Webpack 5 build by removing a reference to a deleted icon
file.

`assets/icons/arrow-down.svg` was deleted in
cfe7891ea4d8d955f7584fbe517fdc684b9c9f2f, but still referenced in CSS
code which causes Webpack 5 to bark. The whole CSS block in which it
was referenced is unused and can be removed.