Fix map scales
The map coloring scales have been totally misleading. Revamp the whole
implementation to make them (hopefully) make more sense.

The scales are now linear from 0% to 200% compared to the Oslo
average (so that 100% = the Oslo average = the middle of the
scale). The (few) values above 200% are simply clamped down to 200%.